Ability to use allow_lazy_user with class methods
The ```allow_lazy_user``` decorator is not designed to be used with class methods that prevents one from using it with class based views. We have added a new decorator ```allow_lazy_user_method``` that accepts the instance/self as required and propagates it through the wrap logic. We have used this version in our project successfully where we needed lazy signup to work with a django rest framework view class.